+++ MoneyManager.Database/Programmability/Procedures/MMSP_DeleteCategory.sql
@@ -1,5 +1,6 @@
 ﻿CREATE PROCEDURE [dbo].[MMSP_DeleteCategory]
-	@Id int
+	@Id int,
+	@UserId int
 AS
-	DELETE FROM [Category] WHERE Id = @Id;
+	DELETE FROM [Category] WHERE Id = @Id and UserId = @UserId;
 RETURN 0
